requirements.txt
Flask==3.0.0
flask-cors==4.0.0
face_recognition==1.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
